requirements.txt
Django==1.11.11
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.7.1
whitenoise==3.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-filter==1.1.0
